try to apply component
