--- WoolLib/feature/main.js
@@ -1,3 +1,0 @@
-import { Login, Start } from "./login.js";
-
-export { Login, Start };
