+++ index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script type="module" src="main.js"></script>
+    <script type="module" src="./WoolLib/index.js"></script>
     <script src="https://code.jquery.com/jquery-3.6.1.min.js"></script>
     <link rel="stylesheet" href="style.css" />
     <title>Document</title>
@@ -15,8 +15,8 @@
     <div class="container">
         <login-screen></login-screen>
     </div>
-    <script src="index.js" type="module"></script>
-    <script type="module" src="main2.js"></script>
+    
+    <script type="module" src="./test/main.js"></script>
 </body>
 
 </html>
