+++ test/instagramVer/insVer.html
@@ -5,6 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://code.jquery.com/jquery-3.6.1.min.js"></script>
+    <script type="module" src="../../WoolLib/index.js"></script>
     <link rel="stylesheet" href="style.css"/>
     <title>Document</title>
 </head>
